Fixed language_tracking cleanup to be more robust if multiple processes are running the same code concurrently.

+++ app/models/language_tracking.rb
@@ -102,7 +102,9 @@ private
       @@last_clean = Time.now
       glob = tag_file('*')
       for file in Dir.glob(glob)
-        if File.mtime(file) < cutoff
+        # I've seen this fail because of files presumably being deleted by
+        # another process between Dir.glob and File.mtime.
+        if File.exist?(file) and File.mtime(file) < cutoff
           File.delete(file)
         end
       end
